✨ feat:auction-list-response

+++ backend/src/main/java/com/auctopus/project/db/domain/AuctionImage.java
@@ -23,7 +23,7 @@ public class AuctionImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int auctionImageSeq;
-    int auctionSeq;
+    Long auctionSeq;
     String imageUrl;
 
 }
